requirements.txt
Flask==3.0.3
requests==2.32.3
psycopg2-binary==2.9.11
